Extract deprecated urls
